highlight ll and ld like int

+++ other/stuff.cpp
@@ -24,6 +24,6 @@ __int128, __float128
 std::decimal::decimal128
 
 // 1e18 < INF < Max_Value / 2
-constexpr long long INF = 0x3FFF'FFFF'FFFF'FFFFll;
+constexpr ll INF = 0x3FFF'FFFF'FFFF'FFFFll;
 // 1e9 < INF < Max_Value / 2
 constexpr int INF = 0x3FFF'FFFF;
